Add support for inception target and protocol
The inception protocol design is separated in 2 classes.
The first is a polling thread to update avatar with the state of the target.
The second implements the command sent to the target (e.g., breakpoints)

This implementation is slower compared to without inception solution (e.g., OpenOCD).
The design flaw lives in how python implements thread with the Global Interpreter Lock (GIL).
By having a thread that polls the inception device in parallel of copying a memory range, the python interpreter adds an overhead when checking locks.